pyproject.toml
[tool.poetry]
name = "poetry-test-project"
version = "0.1.0"
description = ""
authors = ["si0005hp <si0005hp@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry_test_project", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
